// frontend/src/app/api/generated/model/adminUserDto.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRole } from './userRole';


export interface AdminUserDto { 
    id: string;
    username: string;
    role: UserRole;
    enabled: boolean;
}